fixup -Wshorten-64-to-32 warnings
These are emitted by clang on a Ubuntu Trusty amd64 system.
clang version 3.5-1ubuntu1 (trunk) (based on LLVM 3.5)

+++ ext/kgio/poll.c
@@ -160,7 +160,7 @@ retry:
 	}
 	if (nr == 0) return Qnil;
 
-	return poll_result(nr, a);
+	return poll_result((int)nr, a);
 }
 
 /*
